Laid groundwork for collapse
Started the work for the collapsing of the third scene's background slope. Added in the images, animations and animator ccontroller and some varaibles. Will still need to add sprites into the animations and have a way to change the boolean when the new code added in the script is read.